Consolidate Product and Item entities - unified inventory entity model

+++ src/main/java/com/ucop/entity/Item.java
@@ -1,7 +1,16 @@
 package com.ucop.entity;
 
-import jakarta.persistence.*;
 import java.math.BigDecimal;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "items")
@@ -9,7 +18,7 @@ public class Item extends BaseAuditable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false, unique = true, length = 100)
     private String sku;
@@ -45,7 +54,7 @@ public class Item extends BaseAuditable {
 
     // =================== GETTERS/SETTERS ====================
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
